requirements.txt
# Default
fastapi
uvicorn[standard]
python-dotenv